added freshen-list as comman (in setup.py)
